+++ source/Processing.Domain/Customers/Rules/CvrNumberFormatRule.cs
@@ -15,7 +15,7 @@
 using Processing.Domain.Helpers;
 using Processing.Domain.SeedWork;
 
-namespace Processing.Domain.Consumers.Rules
+namespace Processing.Domain.Customers.Rules
 {
     internal class CvrNumberFormatRule : IBusinessRule
     {
