+++ source/Processing.Tests/Domain/BusinessProcesses/MoveIn/ConsumerMoveInTests.cs
@@ -16,7 +16,7 @@ using NodaTime;
 using Processing.Domain.BusinessProcesses.MoveIn;
 using Processing.Domain.BusinessProcesses.MoveIn.Errors;
 using Processing.Domain.Common;
-using Processing.Domain.Consumers;
+using Processing.Domain.Customers;
 using Processing.Domain.EnergySuppliers;
 using Processing.Domain.MeteringPoints;
 using Processing.Domain.MeteringPoints.Events;
