+++ source/Processing.Tests/Domain/MeteringPoints/ChangeOfSupplier/CancellationTests.cs
@@ -14,7 +14,7 @@
 
 using System;
 using NodaTime;
-using Processing.Domain.Consumers;
+using Processing.Domain.Customers;
 using Processing.Domain.EnergySuppliers;
 using Processing.Domain.MeteringPoints;
 using Processing.Domain.MeteringPoints.Events;
